# Optimize RegExpRouter and maybe more