I'm not done yet.

Gonna do more optimizations to the build step as well.