@usualoma 2024-02-28
Hi @aquapi 
Thanks for trying to optimize RegExpRouter.

However, regarding d907da1 and 6dcc1a4, I do not want to merge those changes. I do not see benefit to be gained by merging them.
If you think there are improvements that could be made by merging these PRs, please let us know specifically, e.g. by benchmarking.

But I think that maybe the following changes could have a good outcome.

https://github.com/honojs/hono/pull/2285/commits/d907da1bbb1de166df37ff5c24ec457357fe4cd9#diff-06f981941f95eecd6e249629fd9d5d4 badaf8dfbe96ade6219e67691359df679L217-R212

If there are changes that clearly make a difference in performance, I would like to see them broken down into smaller pieces instead of being grouped into one PR.

Best regards.