@usualoma 2024-02-28
@aquapi

Thanks for the explanation!

I don't think it needs to be a significant improvement. Even a small improvement would be fine, and if the benchmark shows a change, I think we can consider merging. If the benchmark doesn't show a change, then I think we can't merge.