@aquapi 2024-02-28
@usualoma I noticed that Hono has been slower in 3.11 than in 3.8 when I benchmark with my previous framework and Elysia.