@usualoma 2024-02-28
@aquapi Thanks! Your information is very helpful.

The following changes were also observed on the benchmark In particular, `if (typeof staticMatch ! == 'undefined') {` is considered to be an awesome improvement. Could you please make this part a separate PR? I think that could be merged quickly.

<img width="781" alt="CleanShot 2024-02-28 at 22 17 28@2x" src="https://github.com/honojs/hono/assets/30598/dfe30351-01c3-4681-a7b0-4321a22cd535">