@usualoma 2024-02-28
Ah, it's difficult, again, I've checked several times and there seems to be little difference. Might be very slightly faster, though.