@usualoma 2024-03-02
Oh, excuse me, when I pasted the `node:jit` result into honojs-2285-bench above, this PR beat it by "1.01x". But it's a very small difference here, and since they switch every run, I think it's safe to assume there's no difference in performance.

The following is the result of another run immediately afterwards.

```
[ec2-user@ip-10-1-0-53 honojs-2285-bench]$ npm run node:jit

> honojs-2285-bench@1.0.0 node:jit
> esbuild --bundle --platform=node bench.mts | node

cpu: Intel(R) Xeon(R) Platinum 8259CL CPU @ 2.50GHz
runtime: node v21.6.2 (x64-linux)

benchmark      time (avg)             (min … max)       p75       p99      p999
------------------------------------------------- -----------------------------
noop          180 ps/iter     (121 ps … 2'426 ns)    131 ps    241 ps   1.57 ns !
noop          160 ps/iter     (121 ps … 3'048 ns)    131 ps    231 ps    463 ps !
noop          144 ps/iter     (121 ps … 1'131 ns)    129 ps    169 ps    302 ps !
noop          172 ps/iter     (121 ps … 2'857 ns)    129 ps    188 ps    674 ps !
noop          182 ps/iter     (121 ps … 3'232 ns)    131 ps    217 ps    673 ps !
noop          183 ps/iter     (121 ps … 1'856 ns)    131 ps    243 ps   3.35 ns !
noop          213 ps/iter     (121 ps … 2'402 ns)    131 ps    215 ps   3.68 ns !
noop          155 ps/iter     (121 ps … 2'102 ns)    131 ps    217 ps    331 ps !
noop          173 ps/iter     (121 ps … 1'931 ns)    131 ps    247 ps    877 ps !
noop          184 ps/iter     (121 ps … 2'713 ns)    130 ps    208 ps    404 ps !
noop          177 ps/iter     (121 ps … 2'480 ns)    130 ps    243 ps    667 ps !
noop          165 ps/iter     (121 ps … 2'449 ns)    131 ps    232 ps    334 ps !
noop          183 ps/iter     (121 ps … 3'128 ns)    130 ps    194 ps   1.17 ns !
noop          156 ps/iter     (121 ps … 2'009 ns)    131 ps    220 ps    322 ps !
noop          165 ps/iter     (121 ps … 1'764 ns)    130 ps    192 ps    545 ps !

• Startup time
------------------------------------------------- -----------------------------
Old        75'710 ns/iter (30'040 ns … 16'463 µs) 64'628 ns    292 µs  6'045 µs
New        62'039 ns/iter (28'687 ns … 15'356 µs) 60'357 ns    228 µs  2'876 µs

summary for Startup time
  New
   1.22x faster than Old

• Performance
------------------------------------------------- -----------------------------
Old           751 ns/iter     (525 ns … 3'241 ns)    749 ns  2'600 ns  3'241 ns
New           760 ns/iter     (532 ns … 3'685 ns)    748 ns  2'695 ns  3'685 ns

summary for Performance
  Old
   1.01x faster than New
```